mod: restructured sumcheck dir
