+++ sumcheck_protocol/src/lib.rs
@@ -1,4 +1,2 @@
-pub mod prover;
-pub mod verifier;
-pub mod protocol;
-pub mod sumcheck_interactive_simulation;
+pub mod basic_sumcheck;
+pub mod gkr_sumcheck;
